refactor/feat: refactor Application to more of a DDD approach, feat: add fluent validation to endpoints for server side validation and manual validation in .razor's

+++ KS_Application/KS.API/Controllers/OfferController.cs
@@ -1,10 +1,12 @@
 using AutoMapper;
-using KS.Application.DTOs.Offer;
-using KS.Application.DTOs.Response;
+using FluentValidation;
+using FluentValidation.Results;
+using KS.Application.Offers;
+using KS.Application.Offers.Create;
+using KS.Application.Offers.Update;
+using KS.Application.Response;
 using KS.Domain.Offers;
 using Microsoft.AspNetCore.Mvc;
-using System.Text.Json;
-using System.Text.Json.Serialization;
 
 namespace KS.API.Controllers
 {
@@ -14,21 +16,27 @@ namespace KS.API.Controllers
     {
         private IOfferRepository _offerRepository;
         private readonly IMapper _mapper;
+        private readonly IValidator<OfferCreateDTO> _createValidator;
+        private readonly IValidator<OfferUpdateDTO> _updateValidator;
 
-        public OfferController(IOfferRepository offerRepository, IMapper mapper)
+        public OfferController(IOfferRepository offerRepository, IMapper mapper, IValidator<OfferUpdateDTO> updateValidator, IValidator<OfferCreateDTO> createValidator)
         {
             _offerRepository = offerRepository;
             _mapper = mapper;
+            _updateValidator = updateValidator;
+            _createValidator = createValidator;
         }
 
         [HttpGet("getAllList")]
-        public async Task<ActionResult<PagedResult<OfferListDTO>>> GetAllList(int page = 1, int pageSize = 10)
+        public async Task<ActionResult<PagedResultDTO<OfferListDTO>>> GetAllList(int page = 1, int pageSize = 10)
         {
+            if (page < 1) return BadRequest(new { error = "page not greter than 0!" });
+
             var offerDTOs = new List<OfferListDTO>();
             var result = await _offerRepository.GetAll(page, pageSize);
             if (result.Items.Count != 0) offerDTOs = _mapper.Map<List<OfferListDTO>>(result.Items);
 
-            return Ok(new PagedResult<OfferListDTO>
+            return Ok(new PagedResultDTO<OfferListDTO>
             {
                 Data = offerDTOs,
                 TotalCount = result.TotalCount
@@ -38,6 +46,8 @@ namespace KS.API.Controllers
         [HttpGet("getById/{id}")]
         public async Task<ActionResult<OfferDetailsDTO>> GetById(long id)
         {
+            if (id < 1) return BadRequest(new { error = "Id not greter than 0!" });
+
             var offer = await _offerRepository.GetById(id);
             if (offer == null) return NotFound(null);
 
@@ -48,6 +58,16 @@ namespace KS.API.Controllers
         [HttpPost("create")]
         public async Task<ActionResult> Create([FromBody] OfferCreateDTO offerDTO)
         {
+            ValidationResult validationResult = await _createValidator.ValidateAsync(offerDTO);
+
+            if (!validationResult.IsValid) { 
+            
+                var errors = new List<string>();
+                validationResult.Errors.ForEach(e => errors.Add(e.ErrorMessage));
+
+                return BadRequest(new { errors });
+            }
+            
             var offer = _mapper.Map<Offer>(offerDTO);
             await _offerRepository.Create(offer);
 
@@ -57,6 +77,15 @@ namespace KS.API.Controllers
         [HttpPut("update")]
         public async Task<ActionResult> Update([FromBody] OfferUpdateDTO offerDTO)
         {
+            ValidationResult validationResult = await _updateValidator.ValidateAsync(offerDTO);
+
+            if (!validationResult.IsValid) {
+                var errors = new List<string>();
+                validationResult.Errors.ForEach(e => errors.Add(e.ErrorMessage));
+
+                return BadRequest(new { errors });
+            }
+
             var offer = _mapper.Map<Offer>(offerDTO);
             var result = await _offerRepository.Update(offer);
             if (!result) return NoContent();
@@ -67,6 +96,8 @@ namespace KS.API.Controllers
         [HttpDelete("delete/{id}")]
         public async Task<ActionResult> Delete(long id)
         {
+            if (id < 1) return BadRequest(new { error = "Id not greter than 0!" });
+
             var result = await _offerRepository.Delete(id);
             if (!result) return NoContent();
